Skeleton of the filesystem API and RAM filesystem

+++ include/sys/types.h
@@ -43,7 +43,7 @@ typedef int pid_t;
 typedef unsigned long sigset_t;
 typedef int ssize_t;
 typedef long suseconds_t;
-typedef long long time_t;
+typedef long time_t;
 typedef unsigned short uid_t;
 typedef unsigned long useconds_t;
